Don't enable debugging on JVM running tests
While running tests on the build server, we often see the failure
message about unable to bind to a port since it is already in use.
This CL is an attempt to see if not passing the -Xdebug and -Xjdwp
options to the VM that is launched for running tests will help
avoid the issue.

Change-Id: I36d199fff9571c4367ee9b6bb92b2e763df4c775