Don't suggest making activities/fragments private
If you run Analyze > Inspect Code... on a default Android project,
you'll end up with suggestions that you should make your various
activities and services package protected (since they may not be
referenced from other code in the project). However, you can't do that
- at runtime the various inflaters will throw exceptions if you do.

This CL adds knowledge to the class analyzer in IntelliJ such that
it's aware of the Android context classes and does not suggest
reducing visibility of these.

Change-Id: Ib5e7004dba8b66b6541be87f121195453568822a

Conflicts:
	java/java-impl/src/com/intellij/codeInspection/reference/RefJavaManagerImpl.java